+++ src/co/edu/uco/compuconnect/business/business/imp/CentroInformaticaBusinessImp.java
@@ -6,6 +6,9 @@ import java.util.UUID;
 import co.edu.uco.compuconnect.business.assembler.concrete.CentroInformaticaAssembler;
 import co.edu.uco.compuconnect.business.business.CentroInformaticaBusiness;
 import co.edu.uco.compuconnect.business.domain.CentroInformaticaDomain;
+import co.edu.uco.compuconnect.crosscutting.exceptions.CompuconnectBusinessException;
+import co.edu.uco.compuconnect.crosscutting.utils.Messages.CentroInformaticaBusinessImpMessage;
+import co.edu.uco.compuconnect.crosscutting.utils.UtilUUID;
 import co.edu.uco.compuconnect.data.dao.factory.DAOFactory;
 import co.edu.uco.compuconnect.entities.CentroInformaticaEntity;
 
@@ -18,6 +21,18 @@ public final class CentroInformaticaBusinessImp implements CentroInformaticaBusi
 	}
 	@Override
 	public final void crear(final CentroInformaticaDomain datosCentroInformatica) {
+		
+		UUID identificador;
+		CentroInformaticaEntity entityTmp;
+		List<CentroInformaticaEntity> result;
+		
+		do {
+			identificador = UtilUUID.generateNewUUID();
+			entityTmp = CentroInformaticaEntity.create().setIdentificador(identificador);
+			result = daoFactory.getCentroInformaticaDAO().read(entityTmp);
+		}while(!result.isEmpty());
+		
+		
 		daoFactory.getCentroInformaticaDAO().create(CentroInformaticaAssembler.getInstance().toEntityFromDomain(datosCentroInformatica));
 		
 		
@@ -32,12 +47,33 @@ public final class CentroInformaticaBusinessImp implements CentroInformaticaBusi
 
 	@Override
 	public final void modificar(final CentroInformaticaDomain datosModCentroInformatica) {
+		
+		CentroInformaticaEntity entityTmpIdentificador;
+		List<CentroInformaticaEntity> resultIdentificador;
+		entityTmpIdentificador = CentroInformaticaEntity.create().setIdentificador(datosModCentroInformatica.getIdentificador());
+		
+		resultIdentificador = daoFactory.getCentroInformaticaDAO().read(entityTmpIdentificador);
+		
+		if(resultIdentificador.isEmpty()) {
+			throw CompuconnectBusinessException.create(CentroInformaticaBusinessImpMessage.BUSINESS_RULE_CENTRO_INFORMATICA_ID_NOT_FOUND);
+		}
+		
 		daoFactory.getCentroInformaticaDAO().update(CentroInformaticaAssembler.getInstance().toEntityFromDomain(datosModCentroInformatica));
 		
 	}
 
 	@Override
 	public final void eliminar(final CentroInformaticaDomain identificadorEliminacionCentroInformatica) {
+		
+		CentroInformaticaEntity entityTmpIdentificador;
+		List<CentroInformaticaEntity> resultIdentificador;
+		entityTmpIdentificador = CentroInformaticaEntity.create().setIdentificador(identificadorEliminacionCentroInformatica.getIdentificador());
+		
+		resultIdentificador = daoFactory.getCentroInformaticaDAO().read(entityTmpIdentificador);
+		
+		if(resultIdentificador.isEmpty()) {
+			throw CompuconnectBusinessException.create(CentroInformaticaBusinessImpMessage.BUSINESS_RULE_CENTRO_INFORMATICA_ID_NOT_FOUND);
+		}
 		daoFactory.getCentroInformaticaDAO().delete(CentroInformaticaAssembler.getInstance().toEntityFromDomain(identificadorEliminacionCentroInformatica));
 		
 	}
